refactor(internal/librarian/gcloud): rename gcloud package to surfer
Currently, librarian generate for gcloud generates YAML files. Rename
this to internal/librarian/gcloud to free up the gcloud namespace, and
use the surfer namespace to focus on generating YAML files.
