+++ internal/config/language.go
@@ -31,8 +31,8 @@ const (
 	LanguageDotnet = "dotnet"
 	// LanguageFake is the language identifier for Fakes.
 	LanguageFake = "fake"
-	// LanguageGcloud is the language identifier for gcloud command surfaces.
-	LanguageGcloud = "gcloud"
+	// LanguageSurfer is the language identifier for gcloud command surfaces.
+	LanguageSurfer = "gcloud"
 	// LanguageGo is the language identifier for Go.
 	LanguageGo = "go"
 	// LanguageJava is the language identifier for Java.
