+++ internal/librarian/generate_test.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/googleapis/librarian/internal/config"
-	"github.com/googleapis/librarian/internal/librarian/gcloud"
+	"github.com/googleapis/librarian/internal/librarian/surfer"
 	"github.com/googleapis/librarian/internal/sample"
 	"github.com/googleapis/librarian/internal/yaml"
 )
@@ -356,7 +356,7 @@ libraries:
 	}
 }
 
-func TestGenerate_Gcloud(t *testing.T) {
+func TestGenerate_Surfer(t *testing.T) {
 	tempDir := t.TempDir()
 	t.Chdir(tempDir)
 
@@ -386,8 +386,8 @@ libraries:
 	if errors.Is(err, errUnsupportedLanguage) {
 		t.Errorf("expected gcloud to be supported, but got: %v", err)
 	}
-	if !errors.Is(err, gcloud.ErrNoProtosFound) {
-		t.Errorf("expected error %v, got: %v", gcloud.ErrNoProtosFound, err)
+	if !errors.Is(err, surfer.ErrNoProtosFound) {
+		t.Errorf("expected error %v, got: %v", surfer.ErrNoProtosFound, err)
 	}
 }
