+++ internal/librarian/surfer/testdata/publicca/publicca/external_account_keys/_partials/_create_ga.yaml
