+++ internal/librarian/surfer/testdata/publicca/publicca/external_account_keys/create.yaml
